requirements.txt
httpx==0.27.2
httpx[http2]
PyExecJS==1.5.1
pytz==2024.1